+++ src/main/java/javax/util/streamex/DoubleCollector.java
@@ -322,8 +322,9 @@ public interface DoubleCollector<A, R> extends MergingCollector<Double, A, R> {
      *            collector
      * @return a collector which performs the action of the downstream
      *         collector, followed by an additional finishing step
-     * @see #andThen(Function)
+     * @deprecated use {@link #andThen(Function)}.
      */
+    @Deprecated
     static <A, R, RR> DoubleCollector<A, RR> collectingAndThen(DoubleCollector<A, R> downstream,
             Function<R, RR> finisher) {
         return downstream.andThen(finisher);