+++ src/main/java/javax/util/streamex/DoubleStreamEx.java
@@ -307,6 +307,27 @@ public class DoubleStreamEx implements DoubleStream {
         return strategy().newDoubleStreamEx(stream.sorted());
     }
 
+    /**
+     * Returns a stream consisting of the elements of this stream sorted
+     * according to the given comparator. Stream elements are boxed before
+     * passing to the comparator.
+     *
+     * <p>
+     * For ordered streams, the sort is stable. For unordered streams, no
+     * stability guarantees are made.
+     *
+     * <p>
+     * This is a <a href="package-summary.html#StreamOps">stateful intermediate
+     * operation</a>.
+     * 
+     * @param comparator
+     *            a <a
+     *            href="package-summary.html#NonInterference">non-interfering
+     *            </a>, <a
+     *            href="package-summary.html#Statelessness">stateless</a>
+     *            {@code Comparator} to be used to compare stream elements
+     * @return the new stream
+     */
     public DoubleStreamEx sorted(Comparator<Double> comparator) {
         return strategy().newDoubleStreamEx(stream.boxed().sorted(comparator).mapToDouble(Double::doubleValue));
     }