+++ CHANGES.md
@@ -8,6 +8,7 @@
 * Added `StreamEx.toFlatCollection` and `StreamEx.toFlatList` terminal operations.
 * Added `StreamEx.ofSubLists(list, length, shift)` static method.
 * Methods `IntCollector/LongCollector/DoubleCollector.collectingAndThen` are declared as deprecated and may be removed in future releases!
+* Updated documentation
 
 ### 0.3.6